Small bugfixes. Simplify distributor code.
