+++ mava/components/building/__init__.py
@@ -21,10 +21,11 @@ from mava.components.building.adders import (
     ParallelSequenceAdder,
     ParallelSequenceAdderSignature,
 )
-from mava.components.building.dataset_iterators import DatasetIterator
+from mava.components.building.datasets import Dataset
+from mava.components.building.distributor import Distributor
 from mava.components.building.executors import Executor
 from mava.components.building.rate_limiters import OffPolicyRateLimiter
-from mava.components.building.replay_tables import OffPolicyReplayTables
-from mava.components.building.setup import GeneralSetup
+from mava.components.building.setup import SystemSetup
+from mava.components.building.tables import OffPolicyReplayTables
 from mava.components.building.trainers import Trainer
 from mava.components.building.variable_source import VariableSource
